+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html
   - Removed soft Correlation Coefficient
   - Custom accuracy none rounding driven custom diff check
 - Added candles_by_indicator method to find correct Candles list in Hexital
+- Updated Hexital Append with optional 'timeframe' to select specific Candle's to append too
 - Added Movements
   - Added flipped, to detect changes in bool's or solid values
 - Added Indicators